new habit form works... again
