+++ api/database.json
@@ -72,6 +72,45 @@
       "reward": "test",
       "frequency": 6,
       "id": 4
+    },
+    {
+      "id": 5
+    },
+    {
+      "id": 6
+    },
+    {
+      "userId": 7,
+      "habit": "Walk the dog",
+      "cue": "test",
+      "reward": "test",
+      "frequency": "21",
+      "id": 7
+    },
+    {
+      "userId": 7,
+      "habit": "Wash dishes",
+      "cue": "after eating a meal",
+      "reward": "having a clean kitchen",
+      "frequency": "21",
+      "id": 8
+    },
+    {
+      "userId": 7,
+      "habit": "Walk the dog",
+      "cue": "end of day",
+      "reward": "happy puppers",
+      "frequency": "6",
+      "id": 9
+    },
+    {
+      "userId": 7,
+      "habit": "Wash dishes",
+      "cue": "after eating a meal",
+      "reward": "having a clean kitchen",
+      "frequency": "21",
+      "goodHabit": true,
+      "id": 10
     }
   ],
   "count": [
